Add release workflow
This workflow will create a release when a tag is pushed to the repository.
The release will contain the tar.bz2 and zip files of the tools directory.
The tar.bz2 and zip files will also have a corresponding SHA256 file.

Also, printout the size of the tar.bz2 and zip files.